backend/requirements.txt
flask
flask-cors
yfinance
flask-sqlalchemy
psycopg2-binary
python-dotenv
flask-migrate
SQLAlchemy